emotion_recognition/requirements.txt
flask==3.0.2
flask-cors==4.0.0
hsemotion==0.3.0
opencv-python==4.9.0.80
torch==2.2.2
torchvision==0.17.2
torchaudio==2.2.2
Pillow==10.2.0
numpy==1.26.4
timm==0.9.2
onnxruntime==1.17.1
